+++ lib/roles/http/client/client-handshake.c
@@ -846,252 +846,8 @@ html_parser_cb(const hubbub_token *token, void *pw)
 
 #endif
 
-static char *
-lws_strdup(const char *s)
-{
-	char *d = lws_malloc(strlen(s) + 1, "strdup");
-
-	if (d)
-		strcpy(d, s);
-
-	return d;
-}
-
-void
-lws_client_stash_destroy(struct lws *wsi)
-{
-	if (!wsi || !wsi->stash)
-		return;
-
-	lws_free_set_NULL(wsi->stash->address);
-	lws_free_set_NULL(wsi->stash->path);
-	lws_free_set_NULL(wsi->stash->host);
-	lws_free_set_NULL(wsi->stash->origin);
-	lws_free_set_NULL(wsi->stash->protocol);
-	lws_free_set_NULL(wsi->stash->method);
-	lws_free_set_NULL(wsi->stash->iface);
-	lws_free_set_NULL(wsi->stash->alpn);
-
-	lws_free_set_NULL(wsi->stash);
-}
-
-LWS_VISIBLE struct lws *
-lws_client_connect_via_info(struct lws_client_connect_info *i)
-{
-	struct lws *wsi;
-	const struct lws_protocols *p;
-	const char *local = i->protocol;
-
-	if (i->context->requested_kill)
-		return NULL;
-
-	if (!i->context->protocol_init_done)
-		lws_protocol_init(i->context);
-	/*
-	 * If we have .local_protocol_name, use it to select the
-	 * local protocol handler to bind to.  Otherwise use .protocol if
-	 * http[s].
-	 */
-	if (i->local_protocol_name)
-		local = i->local_protocol_name;
-
-	wsi = lws_zalloc(sizeof(struct lws), "client wsi");
-	if (wsi == NULL)
-		goto bail;
-
-	wsi->context = i->context;
-#if defined(LWS_ROLE_H1)
-	/* assert the mode and union status (hdr) clearly */
-	lws_role_transition(wsi, LWSIFR_CLIENT, LRS_UNCONNECTED, &role_ops_h1);
-#else
-	lws_role_transition(wsi, LWSIFR_CLIENT, LRS_UNCONNECTED, &role_ops_raw_skt);
-#endif
-	wsi->desc.sockfd = LWS_SOCK_INVALID;
-
-	/* 1) fill up the wsi with stuff from the connect_info as far as it
-	 * can go.  It's because not only is our connection async, we might
-	 * not even be able to get ahold of an ah at this point.
-	 */
-
-	if (!i->method) /* ie, ws */
-#if defined(LWS_ROLE_WS)
-		if (lws_create_client_ws_object(i, wsi))
-			return NULL;
-#else
-		return NULL;
-#endif
-
-	wsi->user_space = NULL;
-	wsi->pending_timeout = NO_PENDING_TIMEOUT;
-	wsi->position_in_fds_table = LWS_NO_FDS_POS;
-	wsi->c_port = i->port;
-	wsi->vhost = i->vhost;
-	if (!wsi->vhost)
-		wsi->vhost = i->context->vhost_list;
-
-	if (!wsi->vhost) {
-		lwsl_err("At least one vhost in the context is required\n");
-
-		goto bail;
-	}
-
-	wsi->protocol = &wsi->vhost->protocols[0];
-	wsi->client_pipeline = !!(i->ssl_connection & LCCSCF_PIPELINE);
-
-	/* reasonable place to start */
-	lws_role_transition(wsi, LWSIFR_CLIENT, LRS_UNCONNECTED,
-#if defined(LWS_ROLE_H1)
-			&role_ops_h1);
-#else
-	&role_ops_raw_skt);
-#endif
-
-	/*
-	 * 1) for http[s] connection, allow protocol selection by name
-	 * 2) for ws[s], if local_protocol_name given also use it for
-	 *    local protocol binding... this defeats the server
-	 *    protocol negotiation if so
-	 *
-	 * Otherwise leave at protocols[0]... the server will tell us
-	 * which protocol we are associated with since we can give it a
-	 * list.
-	 */
-	if (/*(i->method || i->local_protocol_name) && */local) {
-		lwsl_info("binding to %s\n", local);
-		p = lws_vhost_name_to_protocol(wsi->vhost, local);
-		if (p)
-			wsi->protocol = p;
-	}
-
-	if (wsi && !wsi->user_space && i->userdata) {
-		wsi->user_space_externally_allocated = 1;
-		wsi->user_space = i->userdata;
-	} else
-		/* if we stay in http, we can assign the user space now,
-		 * otherwise do it after the protocol negotiated
-		 */
-		if (i->method)
-			if (lws_ensure_user_space(wsi))
-				goto bail;
-
-#if defined(LWS_WITH_TLS)
-	wsi->tls.use_ssl = i->ssl_connection;
-#else
-	if (i->ssl_connection & LCCSCF_USE_SSL) {
-		lwsl_err("libwebsockets not configured for ssl\n");
-		goto bail;
-	}
-#endif
-
-	/* 2) stash the things from connect_info that we can't process without
-	 * an ah.  Because if no ah, we will go on the ah waiting list and
-	 * process those things later (after the connect_info and maybe the
-	 * things pointed to have gone out of scope.
-	 */
-
-	wsi->stash = lws_zalloc(sizeof(*wsi->stash), "client stash");
-	if (!wsi->stash) {
-		lwsl_err("%s: OOM\n", __func__);
-		goto bail1;
-	}
-
-	wsi->stash->address = lws_strdup(i->address);
-	wsi->stash->path = lws_strdup(i->path);
-	wsi->stash->host = lws_strdup(i->host);
-
-	if (!wsi->stash->address || !wsi->stash->path || !wsi->stash->host)
-		goto bail1;
-
-	if (i->origin) {
-		wsi->stash->origin = lws_strdup(i->origin);
-		if (!wsi->stash->origin)
-			goto bail1;
-	}
-	if (i->protocol) {
-		wsi->stash->protocol = lws_strdup(i->protocol);
-		if (!wsi->stash->protocol)
-			goto bail1;
-	}
-	if (i->method) {
-		wsi->stash->method = lws_strdup(i->method);
-		if (!wsi->stash->method)
-			goto bail1;
-	}
-	if (i->iface) {
-		wsi->stash->iface = lws_strdup(i->iface);
-		if (!wsi->stash->iface)
-			goto bail1;
-	}
-	 /*
-	  * For ws, default to http/1.1 only.  If i->alpn is set, defer to
-	  * whatever he has set in there (eg, "h2").
-	  *
-	  * The problem is he has to commit to h2 before he can find out if the
-	  * server has the SETTINGS for ws-over-h2 enabled; if not then ws is
-	  * not possible on that connection.  So we only try it if he
-	  * assertively said to use h2 alpn.
-	  */
-	if (!i->method && !i->alpn) {
-		wsi->stash->alpn = lws_strdup("http/1.1");
-		if (!wsi->stash->alpn)
-			goto bail1;
-	} else
-		if (i->alpn) {
-			wsi->stash->alpn = lws_strdup(i->alpn);
-			if (!wsi->stash->alpn)
-				goto bail1;
-		}
-
-	if (i->pwsi)
-		*i->pwsi = wsi;
-
-#if defined(LWS_ROLE_H1) || defined(LWS_ROLE_H2)
-	/* if we went on the waiting list, no probs just return the wsi
-	 * when we get the ah, now or later, he will call
-	 * lws_client_connect_via_info2() below.
-	 */
-	if (lws_header_table_attach(wsi, 0) < 0) {
-		/*
-		 * if we failed here, the connection is already closed
-		 * and freed.
-		 */
-		goto bail2;
-	}
-
-#endif
-
-	if (i->parent_wsi) {
-		lwsl_info("%s: created child %p of parent %p\n", __func__,
-				wsi, i->parent_wsi);
-		wsi->parent = i->parent_wsi;
-		wsi->sibling_list = i->parent_wsi->child_list;
-		i->parent_wsi->child_list = wsi;
-	}
-#ifdef LWS_WITH_HTTP_PROXY
-	if (i->uri_replace_to)
-		wsi->http.rw = lws_rewrite_create(wsi, html_parser_cb,
-					     i->uri_replace_from,
-					     i->uri_replace_to);
-#endif
-
-	return wsi;
-
-bail1:
-	lws_client_stash_destroy(wsi);
-
-bail:
-	lws_free(wsi);
-#if defined(LWS_ROLE_H1) || defined(LWS_ROLE_H2)
-bail2:
-#endif
-	if (i->pwsi)
-		*i->pwsi = NULL;
-
-	return NULL;
-}
-
 struct lws *
-lws_client_connect_via_info2(struct lws *wsi)
+lws_http_client_connect_via_info2(struct lws *wsi)
 {
 	struct client_info_stash *stash = wsi->stash;
 
@@ -1154,55 +910,6 @@ bail1:
 #endif
 
 	return NULL;
-}
-
-LWS_VISIBLE struct lws *
-lws_client_connect_extended(struct lws_context *context, const char *address,
-			    int port, int ssl_connection, const char *path,
-			    const char *host, const char *origin,
-			    const char *protocol, int ietf_version_or_minus_one,
-			    void *userdata)
-{
-	struct lws_client_connect_info i;
-
-	memset(&i, 0, sizeof(i));
-
-	i.context = context;
-	i.address = address;
-	i.port = port;
-	i.ssl_connection = ssl_connection;
-	i.path = path;
-	i.host = host;
-	i.origin = origin;
-	i.protocol = protocol;
-	i.ietf_version_or_minus_one = ietf_version_or_minus_one;
-	i.userdata = userdata;
-
-	return lws_client_connect_via_info(&i);
-}
-
-LWS_VISIBLE struct lws *
-lws_client_connect(struct lws_context *context, const char *address,
-			    int port, int ssl_connection, const char *path,
-			    const char *host, const char *origin,
-			    const char *protocol, int ietf_version_or_minus_one)
-{
-	struct lws_client_connect_info i;
-
-	memset(&i, 0, sizeof(i));
-
-	i.context = context;
-	i.address = address;
-	i.port = port;
-	i.ssl_connection = ssl_connection;
-	i.path = path;
-	i.host = host;
-	i.origin = origin;
-	i.protocol = protocol;
-	i.ietf_version_or_minus_one = ietf_version_or_minus_one;
-	i.userdata = NULL;
-
-	return lws_client_connect_via_info(&i);
 }
 
 #if defined(LWS_WITH_SOCKS5)
